feat: move utils.zip to utils.compression
