+++ disdrodb/data_transfer/upload_data.py
@@ -5,7 +5,7 @@ from typing import Optional, List
 
 from disdrodb.api.metadata import _read_yaml_file, get_list_metadata, _write_yaml_file
 from disdrodb.utils.zenodo import _create_zenodo_deposition, _upload_file_to_zenodo
-from disdrodb.utils.zip import _zip_dir
+from disdrodb.utils.compression import _zip_dir
 
 
 def click_upload_option(function: object):
